% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{available_datasets}
\alias{available_datasets}
\title{List available datasets}
\usage{
available_datasets()
}
\value{
character vector of available dataset ids for \code{datasource()}
}
\description{
In \code{sooty_files()} the data source files are grouped by \code{Dataset}, this is the
list of unique datasets, values that can be used in \verb{datasource(<name>)}.
}
\examples{
available_datasets()
}
