% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Create a dataset object. Currently WIP}
\arguments{
\item{id}{a dataset label, see 'dataset()@available_datasets' (get, and settable)}

\item{n}{the number of files (objects) comprising the dataset (get, not settable)}

\item{mindate}{the minimum available date for the files}

\item{maxdate}{the maximum available date for the files}

\item{available_datasets}{the names of available datasets, for 'id' (get, not settable)}

\item{curated}{the full set of files (objects) that sooty knows about that belong to datasets (get, not settable)}

\item{source}{the set of files (objects) belonging to this dataset (get, not settable)}
}
\description{
Generates an object whose "@id" property may be set, which then communicates with a dataset of files/objects that
sooty knows about.
}
\details{
Compare 'curated' to 'sooty_files(curated = FALSE)', if its curated sooty knows what dataset it belongs to
}
\examples{
ds <- dataset()
## available dataset names
ds@available_datasets
## set to one of those
ds@id <- "oisst-tif"
ds
ds@source
}
