% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{datasource}
\alias{datasource}
\alias{dataset}
\title{Create a datasource object. A data source provides a list of files that together comprise a dataset.}
\usage{
datasource(id = NA_character_)

dataset(...)
}
\arguments{
\item{id}{a dataset label, see \code{\link[=available_datasets]{available_datasets()}}}

\item{...}{only used by deprecated function, will become defunct}
}
\description{
Generates an object whose \verb{@id} property may be set, which then communicates with a dataset
of files/objects that sooty knows about.
}
\details{
The following properties are available via the \code{@} slot:
\itemize{
\item \code{id} a dataset label, see \code{\link[=available_datasets]{available_datasets()}} (gettable and settable)
\item \code{n} the number of files (objects) comprising the dataset (get only)
\item \code{mindate} the minimum available date for the files (get only)
\item \code{maxdate} the maximum available date for the files (get only)
\item \code{source} the set of files (objects) belonging to this dataset (get only)
}

By default sooty maintains a local cache of the catalogue used to populate these
properties. Set \code{options("sooty.allow.cache" = FALSE)} to use only the bundled
sysdata, or \code{options("sooty.cache.path" = tempdir())} to redirect the cache
directory. See \code{\link[=sooty_cache_info]{sooty_cache_info()}} for details.
}
\note{
This was originally called \code{dataset()} which usage has now been deprecated.
}
\examples{
op <- options("sooty.allow.cache" = FALSE)
## available dataset names
available_datasets()
## set to one of those
ds <- datasource("ghrsst-tif")
options(op)
}
