% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{.datasets}
\alias{.datasets}
\title{SEALEVEL_GLO_PHY_L4', 'NSIDC_SEAICE_PS_S25km', 'NSIDC_SEAICE_PS_N25km', 'oisst-avhrr-v02r01'}
\usage{
.datasets()
}
\description{
SEALEVEL_GLO_PHY_L4', 'NSIDC_SEAICE_PS_S25km', 'NSIDC_SEAICE_PS_N25km', 'oisst-avhrr-v02r01'
}
