% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsidc.R
\name{nsidc25kmSfiles}
\alias{nsidc25kmSfiles}
\alias{nsidc25kmNfiles}
\title{Return the available files for the readnsidc functions}
\usage{
nsidc25kmSfiles()

nsidc25kmNfiles()
}
\value{
dataframe of source files (S3 objects)
}
\description{
The column 'source' is the one to use, this is constructed from the other columns
'protocol', 'Bucket', and 'Key'. This may change, note that the actual source url of the data
originally can be obtained, see examples.
}
\examples{
(files <- nsidc25kmSfiles())

## generate original source urls (beware this design may change)
sprintf("https://\%s", tail(files$Key))

## to read this with a GDAL-capable read do
dsn <- sprintf("/vsicurl/https://\%s", tail(files$Key))
## but please note these require earthdata credentials set (WIP)
## e.g. terra::rast(dsn[1])  ## variable name is, um variable
}
