% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{datasource}
\alias{datasource}
\title{Create a datasource object. A data source can access a list of files that together comprise a dataset.}
\arguments{
\item{id}{a dataset label, see 'datasorce()@available_datasources' (get, and settable)}

\item{n}{the number of files (objects) comprising the dataset (get, not settable)}

\item{mindate}{the minimum available date for the files}

\item{maxdate}{the maximum available date for the files}

\item{source}{the set of files (objects) belonging to this dataset (get, not settable)}
}
\description{
Generates an object whose "@id" property may be set, which then communicates with a dataset of files/objects that
sooty knows about.
}
\details{
Compare 'curated' to 'sooty_files(curated = FALSE)', if its curated sooty knows what dataset it belongs to
}
\examples{
ds <- datasource()  ## FIXME this should validate empty input
## available dataset names
available_datasets()
## set to one of those
ds@id <- "oisst-tif"
ds
ds@source
}
