% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-cache.R
\name{sooty_cache_info}
\alias{sooty_cache_info}
\title{Show sooty cache status}
\usage{
sooty_cache_info()
}
\value{
A data frame (invisibly) with cache details.
}
\description{
Show sooty cache status
}
\examples{
sooty_cache_info()
}
